feat: organize the dashboard using several slots
