+++ src/app/dashboard/@charts/components/remainingBudget.tsx
@@ -32,8 +32,6 @@ interface IRemainingBudgetProps {
 export default function RemainingBudget({
   data: initialData,
 }: IRemainingBudgetProps) {
-  console.log(initialData);
-
   const { context, dispatch } = use(generalContext);
   const selectedExpensesDay = getSelectedExpensesDay(context);
   const shouldReloadExpenses = getTriggerExpensesReload(context);
